+++ packages/spacecat-shared-data-access/src/models/site/audit-config-type.js
@@ -14,12 +14,19 @@ const AuditConfigType = (data = {}) => {
   const state = {
     disabled: data.disabled || false,
     excludedURLs: data.excludedURLs || [],
+    manualOverwrites: data.manualOverwrites || [],
   };
 
   const self = {
     disabled: () => state.disabled,
-    excludedURLs: () => state.excludedURLs,
-
+    getExcludedURLs: () => state.excludedURLs,
+    updateExcludedURLs: (excludeURLs) => {
+      state.excludedURLs = excludeURLs;
+    },
+    getManualOverwrites: () => state.manualOverwrites,
+    updateManualOverwrites: (manualOverwrites) => {
+      state.manualOverwrites = manualOverwrites;
+    },
     updateDisabled: (newValue) => {
       state.disabled = newValue;
     },
@@ -32,13 +39,15 @@ AuditConfigType.fromDynamoItem = (dynamoItem) => {
   const auditConfigTypeData = {
     disabled: dynamoItem.disabled,
     excludedURLs: dynamoItem.excludedURLs,
+    manualOverwrites: dynamoItem.manualOverwrites,
   };
   return AuditConfigType(auditConfigTypeData);
 };
 
 AuditConfigType.toDynamoItem = (auditConfigType) => ({
   disabled: auditConfigType.disabled(),
-  excludedURLs: auditConfigType.excludedURLs(),
+  excludedURLs: auditConfigType.getExcludedURLs(),
+  manualOverwrites: auditConfigType.getManualOverwrites(),
 });
 
 export default AuditConfigType;