+++ packages/spacecat-shared-data-access/test/it/db.test.js
@@ -49,6 +49,7 @@ function checkSite(site) {
   expect(auditConfig.auditsDisabled()).to.be.a('boolean').which.is.false;
   expect(auditConfig.getAuditTypeConfig(AUDIT_TYPE_LHS_MOBILE)).to.be.an('object');
   expect(auditConfig.getAuditTypeConfig(AUDIT_TYPE_LHS_MOBILE).disabled()).to.be.a('boolean').which.is.false;
+  expect(auditConfig.getAuditTypeConfig(AUDIT_TYPE_LHS_MOBILE).getExcludedURLs()).to.be.a('array').which.is.deep.equal(['https://example.com/excluded']);
   expect(auditConfig.getAuditTypeConfig('non-existing-type')).to.be.undefined;
   expect(auditConfig.getAuditTypeConfig('cwv')).to.be.an('object');
   expect(auditConfig.getAuditTypeConfig('cwv').disabled()).to.be.a('boolean').which.is.true;
@@ -82,6 +83,7 @@ function checkSiteTopPage(siteTopPage) {
   expect(siteTopPage.getSiteId()).to.be.a('string');
   expect(siteTopPage.getURL()).to.be.a('string');
   expect(siteTopPage.getTraffic()).to.be.a('number');
+  expect(siteTopPage.getTopKeyword()).to.be.a('string');
   expect(siteTopPage.getSource()).to.be.a('string');
   expect(siteTopPage.getGeo()).to.be.a('string');
   expect(isIsoDate(siteTopPage.getImportedAt())).to.be.true;
@@ -416,7 +418,7 @@ describe('DynamoDB Integration Test', async () => {
       auditConfig: {
         auditsDisabled: false,
         auditTypeConfigs: {
-          'lhs-mobile': { disabled: false },
+          'lhs-mobile': { disabled: false, excludedURLs: ['https://example.com/excluded'] },
           cwv: { disabled: true },
         },
       },
@@ -767,6 +769,7 @@ describe('DynamoDB Integration Test', async () => {
       siteId,
       url: 'https://example12345.com/page-12345',
       traffic: 360420000,
+      topKeyword: 'keyword12345',
       source: 'rum',
       geo: 'au',
       importedAt: new Date().toISOString(),