+++ packages/spacecat-shared-data-access/test/unit/models/site/audit-config-type.test.js
@@ -73,16 +73,30 @@ describe('AuditConfigType Tests', () => {
     });
   });
 
-  describe('excludedURLs Method', () => {
+  describe('updateExcludedURLs Method', () => {
     it('returns the default excludedURLs array', () => {
       const auditConfigType = AuditConfigType();
-      expect(auditConfigType.excludedURLs()).to.be.an('array').that.is.empty;
+      expect(auditConfigType.getExcludedURLs()).to.be.an('array').that.is.empty;
     });
 
     it('returns the specified excludedURLs array', () => {
       const urls = ['http://example.com', 'http://test.com'];
       const auditConfigType = AuditConfigType({ excludedURLs: urls });
-      expect(auditConfigType.excludedURLs()).to.eql(urls);
+      expect(auditConfigType.getExcludedURLs()).to.eql(urls);
+    });
+
+    it('updates the excludedURLs array to an empty array', () => {
+      const urls = ['http://example.com', 'http://test.com'];
+      const auditConfigType = AuditConfigType({ excludedURLs: urls });
+      auditConfigType.updateExcludedURLs([]);
+      expect(auditConfigType.getExcludedURLs()).to.be.an('array').that.is.empty;
+    });
+
+    it('updates the excludedURLs array', () => {
+      const auditConfigType = AuditConfigType();
+      const newURLs = ['http://newexample.com', 'http://newtest.com'];
+      auditConfigType.updateExcludedURLs(newURLs);
+      expect(auditConfigType.getExcludedURLs()).to.eql(newURLs);
     });
   });
 
@@ -91,7 +105,7 @@ describe('AuditConfigType Tests', () => {
       const dynamoItem = { disabled: true, excludedURLs: ['http://example.com'] };
       const typeConfig = AuditConfigType.fromDynamoItem(dynamoItem);
       expect(typeConfig.disabled()).to.be.true;
-      expect(typeConfig.excludedURLs()).to.eql(['http://example.com']);
+      expect(typeConfig.getExcludedURLs()).to.eql(['http://example.com']);
     });
   });
 
@@ -102,6 +116,28 @@ describe('AuditConfigType Tests', () => {
       const dynamoItem = AuditConfigType.toDynamoItem(auditConfigType);
       expect(dynamoItem.disabled).to.be.true;
       expect(dynamoItem.excludedURLs).to.eql(urls);
+    });
+  });
+
+  describe('updateManualOverwrites Method', () => {
+    it('updates the manualOverwrites array', () => {
+      const auditConfigType = AuditConfigType();
+      const newManualOverwrites = [
+        { brokenTargetURL: 'https://broken.co', targetURL: 'https://overwrite.co' },
+        { brokenTargetURL: 'https://broken.link.co', targetURL: 'https://overwrite.link.co' },
+      ];
+      auditConfigType.updateManualOverwrites(newManualOverwrites);
+      expect(auditConfigType.getManualOverwrites()).to.eql(newManualOverwrites);
+    });
+
+    it('updates the manualOverwrites array to an empty array', () => {
+      const manualOverwrites = [
+        { brokenTargetURL: 'https://broken.co', targetURL: 'https://overwrite.co' },
+        { brokenTargetURL: 'https://broken.link.co', targetURL: 'https://overwrite.link.co' },
+      ];
+      const auditConfigType = AuditConfigType({ manualOverwrites });
+      auditConfigType.updateManualOverwrites([]);
+      expect(auditConfigType.getManualOverwrites()).to.be.an('array').that.is.empty;
     });
   });
 });