+++ packages/spacecat-shared-data-access/test/unit/models/site/audit-config.test.js
@@ -16,7 +16,11 @@ import { expect } from 'chai';
 
 import AuditConfig from '../../../../src/models/site/audit-config.js';
 import {
-  AUDIT_TYPE_BROKEN_BACKLINKS, AUDIT_TYPE_ORGANIC_KEYWORDS, AUDIT_TYPE_ORGANIC_TRAFFIC,
+  AUDIT_TYPE_BROKEN_BACKLINKS,
+  AUDIT_TYPE_EXPERIMENTATION_ESS_DAILY,
+  AUDIT_TYPE_EXPERIMENTATION_ESS_MONTHLY,
+  AUDIT_TYPE_ORGANIC_KEYWORDS,
+  AUDIT_TYPE_ORGANIC_TRAFFIC,
 } from '../../../../src/models/audit.js';
 
 describe('AuditConfig Tests', () => {
@@ -27,6 +31,10 @@ describe('AuditConfig Tests', () => {
       const auditTypeConfigs = auditConfig.getAuditTypeConfigs();
       expect(auditTypeConfigs[AUDIT_TYPE_BROKEN_BACKLINKS]).to.be.an('object');
       expect(auditTypeConfigs[AUDIT_TYPE_BROKEN_BACKLINKS].disabled()).to.be.true;
+      expect(auditTypeConfigs[AUDIT_TYPE_EXPERIMENTATION_ESS_DAILY]).to.be.an('object');
+      expect(auditTypeConfigs[AUDIT_TYPE_EXPERIMENTATION_ESS_DAILY].disabled()).to.be.true;
+      expect(auditTypeConfigs[AUDIT_TYPE_EXPERIMENTATION_ESS_MONTHLY]).to.be.an('object');
+      expect(auditTypeConfigs[AUDIT_TYPE_EXPERIMENTATION_ESS_MONTHLY].disabled()).to.be.true;
       expect(auditTypeConfigs[AUDIT_TYPE_ORGANIC_KEYWORDS]).to.be.an('object');
       expect(auditTypeConfigs[AUDIT_TYPE_ORGANIC_KEYWORDS].disabled()).to.be.true;
       expect(auditTypeConfigs[AUDIT_TYPE_ORGANIC_TRAFFIC]).to.be.an('object');